% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfs-methods.R
\name{cfs_s3_dir}
\alias{cfs_s3_dir}
\title{cfs_s3_dir}
\usage{
cfs_s3_dir(
  product,
  path = "neon4cast-drivers/noaa/cfs/",
  endpoint = "https://sdsc.osn.xsede.org",
  bucket = paste0("bio230014-bucket01/", path, product)
)
}
\arguments{
\item{product}{target folder, e.g. "6hrly/00" for cycle 00, 6-hrly product}

\item{path}{path inside bucket}

\item{endpoint}{endpoint url}

\item{bucket}{bucket name}
}
\description{
cfs_s3_dir
}
