% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfs-methods.R
\name{cfs_to_parquet}
\alias{cfs_to_parquet}
\title{cfs_to_parquet}
\usage{
cfs_to_parquet(
  dates = Sys.Date() - 1L,
  path = "cfs_parquet",
  ensemble = cfs_ensemble(),
  bands = cfs_bands(),
  sites = sf::st_shift_longitude(neon_sites()),
  horizon = cfs_horizon,
  all_bands = cfs_all_bands(),
  url_builder = cfs_urls,
  cycle = "00",
  family = "ensemble",
  partitioning = c("reference_datetime", "site_id")
)
}
\arguments{
\item{dates}{a vector of reference_datetimes}

\item{path}{path to local directory or S3 bucket (see \code{\link[arrow:write_dataset]{arrow::write_dataset()}})}

\item{ensemble}{list of ensembles}

\item{bands}{named vector of bands to extract}

\item{sites}{sf object of sites}

\item{horizon}{vector of horizons (in hours, as integer values), or as constructor
function.  CFS requires a dynamic constructor since horizon varies by reference date
and ensemble.}

\item{all_bands}{vector of all band names, needed for
\verb{[gdalcubes::stack_cube()]}}

\item{url_builder}{function that constructs URLs to access grib files.
must be a function of horizon, ens, reference_datetime, cycle, and any
optional additional arguments.}

\item{cycle}{cycle indicating start time when forecast was generated
(i.e. "00", "06", "12", or "18" hours into reference_datetime)}

\item{family}{statistical family, e.g. 'ensemble' or 'normal'}

\item{partitioning}{partitioning structure used in writing the parquet data}
}
\description{
cfs_to_parquet
}
