% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalcubes-methods.R
\name{cube_extract}
\alias{cube_extract}
\title{cube_extract}
\usage{
cube_extract(
  reference_datetime,
  ensemble = gefs_ensemble(),
  horizon = gefs_horizon(),
  sites = neon_sites(),
  bands = gefs_bands(),
  all_bands = gefs_all_bands(),
  url_builder = gefs_urls,
  cycle = "00",
  shm = NULL
)
}
\arguments{
\item{reference_datetime}{date forecast is generated}

\item{ensemble}{vector of ensembles, string}

\item{horizon}{vector of horizons to extract (in hours)}

\item{sites}{sf object of sites to extract}

\item{bands}{numeric vector of bands to extract, see details}

\item{all_bands}{names of all bands in grib file}

\item{url_builder}{helper function to construct remote URLs.}

\item{cycle}{start cycle of forecast}

\item{shm}{file path that can be used to replace remote URLs
if data is mirrored to filesystem}
}
\description{
cube_extract
}
