% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalcubes-to-parquet.R
\name{efi_format_cubeextract}
\alias{efi_format_cubeextract}
\title{efi_format_cubeextract}
\usage{
efi_format_cubeextract(dfs, date, sf_sites = neon_sites())
}
\arguments{
\item{dfs}{list of tables from grib_extract}

\item{date}{forecast reference_datetime}

\item{sites}{sites (as sf object)}
}
\description{
reformat results from grib_extract to EFI format
}
