% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalcubes_cloud_config.R
\name{gdalcubes_cloud_config}
\alias{gdalcubes_cloud_config}
\title{gdalcubes cloud config}
\usage{
gdalcubes_cloud_config()
}
\description{
set the recommended environmental variables for cloud-based access
}
