% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gefs-methods.R
\name{gefs_ensemble}
\alias{gefs_ensemble}
\title{gefs ensemble list}
\usage{
gefs_ensemble()
}
\value{
Generates the strings for the 30 perturbed ensembles and control
If only mean and spread are needed, manually pass
\code{c(mean = "geavg", spr = "gespr")}
}
\description{
gefs ensemble list
}
