% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalcubes-methods.R
\name{gefs_grib_collection}
\alias{gefs_grib_collection}
\title{gefs_grib_collection}
\usage{
gefs_grib_collection(
  ens,
  reference_datetime = Sys.Date(),
  horizon = gefs_horizon(),
  band_names = all_gefs_bands(),
  cycle = "00",
  ...
)
}
\arguments{
\item{ens}{ensemble the ensemble member to process, e.g. 'gepavg'
can also be geavg (mean), gespr (esemble spread),
gec00 (control) or gep01-gep30 (perturbed)}

\item{reference_datetime}{start date for forecast}

\item{horizon}{list of horizon values, in hours (see \code{\link[=gefs_horizon]{gefs_horizon()}})}

\item{cycle}{forecast cycle, GEFS forecasts are produced four times a day,
at 00 (for 35 day horizon) and at 06 ,12, 18 hrs (at 16 day horizon)
Note that the control member, gec00, also has only a 16 day horizon}

\item{...}{additional arguments got create_image_collection}
}
\value{
gdalcubes::image_collection()
}
\description{
Generate grib URLs for a requested forecast timeseries (geospatial data cube)
as a \code{\link[gdalcubes:image_collection]{gdalcubes::image_collection()}} object.
}
\details{
\url{https://www.nco.ncep.noaa.gov/pmb/products/gens/}
}
