% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gefs-methods.R
\name{gefs_horizon}
\alias{gefs_horizon}
\title{gefs_horizon}
\usage{
gefs_horizon(gefs_version = Sys.getenv("GEFS_VERSION", "v12"), ...)
}
\arguments{
\item{gefs_version}{GEFS version: v11 covers reference_datetimes from
Jan 1, 2017 to Sept 24, 2020. v12 covers all dates following that.
(earlier versions are not available on AWS).
Note that gefs v11 has only 20 ensemble members, 16 day horizons, and
all at 6 hour intervals.  gefs v11 also has fewer bands.
export}

\item{...}{additional parameters (not used, for cross-compatibility only)}
}
\value{
list of horizon values (for cycle 00, gepNN forecasts)
}
\description{
gefs_horizon
}
