% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gefs-pseudo-measures.R
\name{gefs_pseudo_measures}
\alias{gefs_pseudo_measures}
\title{gefs_pseudo_measures}
\usage{
gefs_pseudo_measures(
  dates = Sys.Date() - 1L,
  path = "gefs_pseudo",
  ensemble = gefs_ensemble(),
  bands = gefs_bands(),
  sites = neon_sites(),
  horizon = c("003", "006"),
  all_bands = gefs_all_bands(),
  url_builder = gefs_urls,
  cycles = c("00", "06", "12", "18"),
  partitioning = c("reference_datetime")
)
}
\arguments{
\item{dates}{a vector of reference_datetimes}

\item{path}{path to local directory or S3 bucket (see \code{\link[arrow:write_dataset]{arrow::write_dataset()}})}

\item{ensemble}{list of ensembles}

\item{bands}{named vector of bands to extract}

\item{sites}{sf object of sites}

\item{horizon}{vector of horizons (in hours, as integer values), or as constructor
function.  CFS requires a dynamic constructor since horizon varies by reference date
and ensemble.  000 horizon is added automatically}

\item{all_bands}{vector of all band names, needed for
\verb{[gdalcubes::stack_cube()]}}

\item{url_builder}{function that constructs URLs to access grib files.
must be a function of horizon, ens, reference_datetime, cycle, and any
optional additional arguments.}

\item{partitioning}{partitioning structure used in writing the parquet data}

\item{cycle}{cycle indicating start time when forecast was generated
(i.e. "00", "06", "12", or "18" hours into reference_datetime)}
}
\description{
gefs_pseudo_measures
}
