% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gefs-methods.R
\name{gefs_s3_dir}
\alias{gefs_s3_dir}
\title{gefs_s3_dir}
\usage{
gefs_s3_dir(
  product = "stage1",
  path = "neon4cast-drivers/noaa/",
  gefs_version = Sys.getenv("GEFS_VERSION", "v12"),
  endpoint = "https://sdsc.osn.xsede.org",
  bucket = "bio230014-bucket01"
)
}
\arguments{
\item{product}{product code, e.g. stage1}

\item{path}{path inside bucket}

\item{gefs_version}{gefs-version (used in path)}

\item{endpoint}{endpoint url}

\item{bucket}{bucket name}
}
\value{
s3 bucket object (an arrow S3 SubTreeFileSystem object)
}
\description{
Helper function that returns a path to the EFI S3 directory
}
\examples{
\donttest{
gefs_s3_dir()
}
}
