% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gefs-methods.R
\name{gefs_to_parquet}
\alias{gefs_to_parquet}
\title{gefs_to_parquet}
\usage{
gefs_to_parquet(
  dates = Sys.Date() - 1L,
  path = "gefs_parquet",
  ensemble = gefs_ensemble(),
  bands = gefs_bands(),
  sites = neon_sites(),
  horizon = gefs_horizon(),
  all_bands = gefs_all_bands(),
  url_builder = gefs_urls,
  cycle = "00",
  partitioning = c("reference_datetime", "site_id")
)
}
\arguments{
\item{dates}{a vector of reference_datetimes}

\item{path}{path to local directory or S3 bucket (see \code{\link[arrow:write_dataset]{arrow::write_dataset()}})}

\item{ensemble}{list of ensembles}

\item{bands}{named vector of bands to extract}

\item{sites}{sf object of sites}

\item{horizon}{vector of horizons (in hours, as integer values), or as constructor
function.  CFS requires a dynamic constructor since horizon varies by reference date
and ensemble.}

\item{all_bands}{vector of all band names, needed for
\verb{[gdalcubes::stack_cube()]}}

\item{url_builder}{function that constructs URLs to access grib files.
must be a function of horizon, ens, reference_datetime, cycle, and any
optional additional arguments.}

\item{cycle}{cycle indicating start time when forecast was generated
(i.e. "00", "06", "12", or "18" hours into reference_datetime)}

\item{partitioning}{partitioning structure used in writing the parquet data}
}
\description{
gefs_to_parquet
}
