% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gefs-methods.R
\name{gefs_urls}
\alias{gefs_urls}
\title{gefs_urls}
\usage{
gefs_urls(
  ens = "geavg",
  reference_datetime = Sys.Date(),
  horizon = gefs_horizon(),
  cycle = "00",
  series = "atmos",
  gefs_version = Sys.getenv("GEFS_VERSION", "v12"),
  base = "https://noaa-gefs-pds.s3.amazonaws.com"
)
}
\arguments{
\item{ens}{ensemble member for which URLs should be generated}

\item{reference_datetime}{date forecast is produced}

\item{horizon}{vector of horizons (in hours, as integer values), or as constructor
function.  CFS requires a dynamic constructor since horizon varies by reference date
and ensemble.}

\item{cycle}{cycle indicating start time when forecast was generated
(i.e. "00", "06", "12", or "18" hours into reference_datetime)}

\item{series}{data series (used only by gefs_v12)}

\item{gefs_version}{GEFS version: v11 covers reference_datetimes from
Jan 1, 2017 to Sept 24, 2020. v12 covers all dates following that.
(earlier versions are not available on AWS).
Note that gefs v11 has only 20 ensemble members, 16 day horizons, and
all at 6 hour intervals.  gefs v11 also has fewer bands.
export}

\item{base}{NOAA GEFS AWS Bucket}
}
\description{
gefs_urls
}
