% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalcubes-methods.R
\name{gefs_view}
\alias{gefs_view}
\title{gefs_view}
\usage{
gefs_view(
  t0 = Sys.Date(),
  t1 = t0 + 35L - lubridate::seconds(1),
  box = gefs_bbox(),
  dx = 0.5,
  dy = 0.5,
  dt = "PT3H",
  crs = "EPSG:4326",
  ...
)
}
\arguments{
\item{t0}{start time (Date or datetime object)}

\item{t1}{end time. 35 days after start by default.}

\item{box}{bounding box giving (xmin, ymin, xmax, ymax) numeric vector
(or an sf_bbox object)}

\item{dx}{resolution, x-coordinate}

\item{dy}{resolution, y-coordinate}

\item{dt}{resolution, time coordinate, ISO8601 temporal-duration notation}

\item{crs}{Coordinate reference system (string)}

\item{...}{additional options, see \code{\link[gdalcubes:cube_view]{gdalcubes::cube_view()}}}
}
\value{
a cube_view object
}
\description{
A thin wrapper around gdalcubes::cube_view
with the default configuration for GEFS
}
\examples{
view <- gefs_view(as.Date("2022-12-01"))

}
