% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_gefs.R
\name{noaa_gefs}
\alias{noaa_gefs}
\title{noaa_gefs}
\usage{
noaa_gefs(
  date = Sys.Date(),
  cycle = "00",
  ...,
  s3 = arrow::s3_bucket("drivers", endpoint_override = "data.ecoforecast.org")
)
}
\arguments{
\item{date}{reference_datetime}

\item{cycle}{start hour of reference datetime}

\item{s3}{s3 bucket or local path}
}
\description{
Compatibility wrapper for noaa gefs stage1 access
}
