% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfs_forecast.R
\name{cfs_to_parquet}
\alias{cfs_to_parquet}
\title{cfs_to_parquet}
\usage{
cfs_to_parquet(
  dates = Sys.Date() - 1L,
  path = "gefs_parquet",
  ensemble = cfs_ensemble(),
  bands = cfs_bands(),
  sites = neon_sites(),
  horizon = cfs_horizon(),
  all_bands = cfs_all_bands(),
  url_builder = cfs_urls,
  cycle = "00",
  partitioning = c("reference_datetime", "site_id")
)
}
\arguments{
\item{dates}{a vector of reference_datetimes}

\item{path}{path to local directory or S3 bucket (see \code{\link[arrow:write_dataset]{arrow::write_dataset()}})}

\item{ensemble}{vector of ensemble values (e.g. 'gep01', 'gep02', ...)}

\item{bands}{numeric vector of bands to extract, see details}

\item{sites}{sf object of sites to extract}

\item{horizon}{vector of horizons to extract (in hours)}

\item{all_bands}{names of all bands in grib file}

\item{url_builder}{helper function to construct remote URLs.
Must be a function of ens, reference_datetime, horizon, cycle, and
any named arguments passed via ...}

\item{partitioning}{partitioning structure used in writing the parquet data}
}
\description{
cfs_to_parquet
}
