% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalcubes-to-parquet.R
\name{efi_format_cubeextract}
\alias{efi_format_cubeextract}
\title{efi_format_cubeextract}
\usage{
efi_format_cubeextract(
  dfs,
  reference_datetime,
  sites = neon_sites(),
  bands = gefs_bands()
)
}
\arguments{
\item{dfs}{list of tables from grib_extract}

\item{sites}{sites (as sf object)}

\item{bands}{named vector of band numbers, see \code{\link[=gefs_bands]{gefs_bands()}}}

\item{date}{forecast reference_datetime}
}
\description{
reformat results from grib_extract to EFI format
}
