% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gefs-methods.R
\name{gefs_bands}
\alias{gefs_bands}
\title{mapping of gefs_bands to variable names}
\usage{
gefs_bands(zero_horizon = FALSE)
}
\arguments{
\item{zero_horizon}{GEFS zero-horizon data uses different band numbering,
so this must be set to TRUE if zero-horizon data is desired.
export}
}
\description{
mapping of gefs_bands to variable names
}
