% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gefs-methods.R
\name{gefs_horizon}
\alias{gefs_horizon}
\title{gefs_horizon}
\usage{
gefs_horizon(...)
}
\arguments{
\item{...}{additional parameters (not used, for cross-compatibility only)}
}
\value{
list of horizon values (for cycle 00, gepNN forecasts)
}
\description{
gefs_horizon
}
