% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalcubes-to-parquet.R
\name{gefs_s3_dir}
\alias{gefs_s3_dir}
\title{gefs_s3_dir}
\usage{
gefs_s3_dir(
  product,
  path = "neon4cast-drivers/noaa/gefs-v12/",
  endpoint = "https://sdsc.osn.xsede.org",
  bucket = paste0("bio230014-bucket01/", path, product)
)
}
\arguments{
\item{product}{product code, e.g. stage1}

\item{path}{path inside bucket}

\item{endpoint}{endpoint url}

\item{bucket}{bucket name}
}
\value{
s3 bucket object (an arrow S3 SubTreeFileSystem object)
}
\description{
Helper function that returns a path to the EFI S3 directory
}
