% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalcubes-to-parquet.R
\name{gefs_to_parquet}
\alias{gefs_to_parquet}
\title{gefs_to_parquet}
\usage{
gefs_to_parquet(
  dates,
  ensemble,
  s3_dir,
  sites = neon_sites(),
  bands = gefs_bands(),
  cycle = "00",
  horizon = gefs_horizon()
)
}
\arguments{
\item{dates}{a vector of reference_datetimes}

\item{sites}{sf object of sites to extract}

\item{bands}{numeric vector of bands to extract, see details}

\item{cycle}{forecast cycle, GEFS forecasts are produced four times a day,
at 00 (for 35 day horizon) and at 06 ,12, 18 hrs (at 16 day horizon)
Note that the control member, gec00, also has only a 16 day horizon}

\item{horizon}{list of horizon values, in hours (see \code{\link[=gefs_horizon]{gefs_horizon()}})}
}
\description{
gefs_to_parquet
}
