% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalcubes-to-parquet.R
\name{gefs_to_parquet}
\alias{gefs_to_parquet}
\title{gefs_to_parquet}
\usage{
gefs_to_parquet(
  dates = Sys.Date() - 1L,
  path = "gefs_parquet",
  ensemble = gefs_ensemble(),
  sites = neon_sites(),
  bands = gefs_bands(),
  cycle = "00",
  horizon = gefs_horizon(),
  partitioning = c("reference_datetime", "site_id")
)
}
\arguments{
\item{dates}{a vector of reference_datetimes}

\item{path}{path to local directory or S3 bucket (see \code{\link[arrow:write_dataset]{arrow::write_dataset()}})}

\item{ensemble}{vector of ensemble values (e.g. 'gep01', 'gep02', ...)}

\item{sites}{sf object of sites to extract}

\item{bands}{numeric vector of bands to extract, see details}

\item{cycle}{forecast cycle, GEFS forecasts are produced four times a day,
at 00 (for 35 day horizon) and at 06 ,12, 18 hrs (at 16 day horizon)
Note that the control member, gec00, also has only a 16 day horizon}

\item{horizon}{list of horizon values, in hours (see \code{\link[=gefs_horizon]{gefs_horizon()}})}

\item{partitioning}{partitioning structure used in writing the parquet data}
}
\description{
gefs_to_parquet
}
