% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalcubes-methods.R
\name{grib_extract}
\alias{grib_extract}
\title{grib_extract}
\usage{
grib_extract(
  ens,
  reference_datetime = Sys.Date(),
  bands = gefs_bands(),
  sites = neon_sites(),
  horizon = gefs_horizon(),
  all_bands = gefs_all_bands(),
  url_builder = gefs_urls,
  cycle = "00",
  ...
)
}
\arguments{
\item{bands}{numeric vector of bands to extract, see details}

\item{sites}{sf object of sites to extract}

\item{horizon}{vector of horizons to extract (in hours)}

\item{all_bands}{names of all bands in grib file}

\item{url_builder}{helper function to construct remote URLs.
Must be a function of ens, reference_datetime, horizon, cycle, and
any named arguments passed via ...}
}
\value{
gdalcubes::image_collection()
}
\description{
grib_extract
}
\details{
See \url{https://www.nco.ncep.noaa.gov/pmb/products/gens/} for
details on GEFS data, including horizon, cycle, and band information.
This function is a simple wrapper around
\code{\link[gdalcubes:extract_geom]{gdalcubes::extract_geom()}}. .
}
