% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalcubes-methods.R
\name{neon_sites}
\alias{neon_sites}
\title{NEON sites as a simple features point geometry object}
\usage{
neon_sites()
}
\value{
an sf object with coordinates for NEON sites
}
\description{
NEON sites as a simple features point geometry object
}
