% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gefs.R
\name{noaa_gefs}
\alias{noaa_gefs}
\title{Stream NOAA GEFS GRIB to parquet}
\usage{
noaa_gefs(
  date = Sys.Date(),
  cycle = "00",
  threads = 70,
  gdal_ops = "",
  s3 = arrow::s3_bucket("drivers", endpoint_override = "data.ecoforecast.org"),
  max_horizon = 840,
  purge = TRUE,
  quiet = FALSE,
  dest = ".",
  locations = paste0("https://github.com/eco4cast/neon4cast-noaa-download/",
    "raw/master/noaa_download_site_list.csv"),
  name_pattern = "noaa/gefs-v12/stage1/{cycle_int}/{nice_date}/part-0.parquet"
)
}
\arguments{
\item{date}{date forecast is issued}

\item{cycle}{hour forecast is issued (00, 06, 12, 18)}

\item{threads}{parallel processes to run}

\item{gdal_ops}{options to GDAL (e.g. compression)}

\item{s3}{an S3 bucket address, from \link[arrow:s3_bucket]{arrow::s3_bucket}}

\item{max_horizon}{maximum horizon}

\item{purge}{logical, clear downloaded/converted tif?}

\item{quiet}{logical, verbose output?}

\item{dest}{destination directory}

\item{locations}{where to access a list of download sites (lat/long)}

\item{name_pattern}{Naming pattern for upload bucket (glue format)}
}
\value{
invisibly, the processx log
}
\description{
Stream NOAA GEFS GRIB to parquet
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
noaa_gefs()
\dontshow{\}) # examplesIf}
}
